requirements.txt
google-api-core
google-cloud-bigquery
google-cloud-bigquery[pandas]
google-cloud-core
cx-Oracle==8.3.0
pydash==5.0.2
sqlalchemy==1.4.36
rq == 1.10.1
pandas == 1.3.5
fsspec
gcsfs
tqdm==4.64.0
cPython==0.0.6